requirements.txt
Django==4.2.7
django-bootstrap-v5
markdown==3.5.1
pygments==2.17.2
django-mdeditor==0.1.20